Used scheduler when preloading
